+++ debug/site.js
@@ -10,7 +10,6 @@ var map = new mapboxgl.Map({
 });
 
 map.addControl(new mapboxgl.Navigation());
-map.addControl(new mapboxgl.Attribution());
 
 var geojson = new mapboxgl.GeoJSONSource({
     data: {